Restructure of server side. Created handler, service folder and spread out middleware file
