+++ server/docker-compose.yml
@@ -1,9 +1,7 @@
-version: '3.8'
-
 services:
   postgres:
     image: postgres:latest
-    container_name: postgres
+    container_name: postgres_goBookCrud
     environment:
       POSTGRES_DB: mydatabase
       POSTGRES_USER: myuser
@@ -16,7 +14,7 @@ services:
 
   pgadmin:
     image: dpage/pgadmin4:latest
-    container_name: pgadmin
+    container_name: pgadmin_goBookCrud
     environment:
       PGADMIN_DEFAULT_EMAIL: admin@admin.com
       PGADMIN_DEFAULT_PASSWORD: admin
